create arithmetic module. create script to add files.
