+++ setup.py
@@ -51,7 +51,8 @@ setup_args = {
                 'scripts/dpss_xtalk_filter_run_baseline_parallelized.py',
                 'scripts/dpss_delay_filter_run.py', 'scripts/chunk_cal_files.py',
                 'scripts/chunk_data_files.py', 'scripts/time_average.py',
-                'scripts/time_average_baseline_parallelized.py'],
+                'scripts/time_average_baseline_parallelized.py',
+                'scripts/sum_files.py'],
     'version': version.version,
     'package_data': {'hera_cal': data_files},
     'install_requires': [
